PythonAPI/requirements.txt
Flask
torch
torchvision
Pillow
flask-cors
opencv-python
pytesseract
pdf2image
spacy
easyocr 
pillow 

numpy
